put needsReviewTestConfig into its own file
