+++ src/electron/platform/android/test-configs/android-test-configs.ts
@@ -1,10 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import {
-    automatedChecksTestConfig,
-    needsReviewTestConfig,
-} from 'electron/platform/android/test-configs/automated-checks/test-config';
+import { automatedChecksTestConfig } from 'electron/platform/android/test-configs/automated-checks/test-config';
+import { needsReviewTestConfig } from 'electron/platform/android/test-configs/needs-review/test-config';
 import { TestConfig } from 'electron/types/test-config';
 
 export const androidTestConfigs: TestConfig[] = [automatedChecksTestConfig, needsReviewTestConfig];
